refactor: densify remaining ≥450 cases modules and ratchet ceilings (#4310)
Topic-split seven fat planning case libraries into barrels plus sibling
modules; ratchet planning-cases, expand-peace-pin, and residual-fat-pin
soft ceilings from 500 to 450 with SPEC/manifest updates.

Refs #4310
