+++ packages/colonizethis_ai/test/planning/expand_phase_planner_survival_zero_regiment_peace_guard_cases.dart
@@ -1,0 +1,257 @@
+// Case bodies for `expand_phase_planner_survival_multi_front_peace_test.dart` (Refs #3977 Phase 6).
+// Registered from the thin contract file of the same stem.
+// Pin/row coverage is preserved 1:1 from the former inline suite.
+
+// Pins canonical homes in `expand_phase_planner.dart` for
+// `stalledZeroRegimentAllFactionPeaceTargets`,
+// `stalledZeroRegimentGpPeaceTargets`,
+// `mutualZeroRegimentGpStalematePeaceTargets`,
+// `mutualExhaustedBelowQuotaGpStalematePeaceTargets`, and
+// `multiFrontNonBlockerGpPeaceTargets` (Refs #2509 S1). Also covers
+// `stalledZeroRegimentGpPeaceTargets` and
+// `mutualZeroRegimentGpStalematePeaceTargets` EXPAND-phase zero-regiment
+// survival peace deciders at their new home in `expand_phase_planner.dart`
+// (Refs #2509 S1).
+//
+// Both deciders were relocated from `diplomacy_planner_peace_targets.dart`
+// so they survive the now-completed S1 deletion of that file. The canonical
+// implementations live in `expand_phase_planner.dart`;
+// `diplomacy_planner_peace_targets.dart` previously retained thin delegating stubs
+// for the legacy `diplomacy_planner_below_quota_peace_part3_test.dart`
+// § "all GP wars when stalled" fixture and the in-file
+// `_survivalGreatPowerPeaceTargets` /
+// `collectStalledGreatPowerPeaceTargets` `zeroRegimentBlockerPeace` /
+// `stalledOwExpansionNeedsPeacePass` consumer chains until the planned
+// deletion.
+//
+// Live consumers (post-relocation):
+//   * `stalledZeroRegimentGpPeaceTargets` is the broad EXPAND
+//     zero-regiment rebuild shortcut from
+//     `SPEC/ai/ai-architecture.md` § Diplomacy targeting — "when
+//     stalled below quota with zero regiments, peace every at-war
+//     Great Power so rebuild is not blocked by futile fronts". It
+//     peaces every at-war GP once the active player is inside the
+//     stalled OW band and holds zero standing regiments, sorted
+//     ascending by `factionId` for downstream offer-peace
+//     determinism.
+//   * `mutualZeroRegimentGpStalematePeaceTargets` is the sole-GP
+//     mutual-stalemate carve-out. It peaces the lone GP enemy when
+//     both sides have zero standing regiments and the active player
+//     is stalled — the only path that exits an army-exhausted GP-only
+//     frontier where the broader `stalledZeroRegimentGpPeaceTargets`
+//     arm is overridden by the `collectStalledGreatPowerPeaceTargets`
+//     GP-only-frontier carve-out re-adding the canonical OW frontier
+//     blocker to the keep-at-war set.
+//
+// Sibling test coverage that this file complements (but does not duplicate):
+//
+//   * `diplomacy_planner_below_quota_peace_part3_test.dart` § "all GP
+//     wars when stalled" exercises the broader arm through the legacy
+//     `diplomacy_planner_peace_targets.dart` entry point and pins the
+//     "filters minors out of GP results" invariant. Both legacy
+//     fixtures depend on the delegating stubs and continue to pass
+//     unchanged after the canonical bodies relocated here.
+//   * `domain_planner_orchestrator_*_two_gp_peace_test.dart` exercise
+//     the deciders through the orchestrator's `runDiplomacyPlanner`
+//     pass under EXPAND / COLONIAL / DEVELOP phases. Those flows rely
+//     on the same post-delegation return values pinned here.
+//
+// Behavioral invariants pinned at the canonical entry points:
+//
+//   1. `stalledZeroRegimentGpPeaceTargets` short-circuits to `const []`
+//      when the active player's `oldWorldProvincesOwned` exceeds
+//      `kStalledOldWorldProvinceThreshold` — outside the stalled band
+//      the rebuild-peace arm does not engage.
+//   2. `stalledZeroRegimentGpPeaceTargets` short-circuits to `const []`
+//      when the active player still holds at least one standing
+//      regiment — the rebuild-peace arm is a zero-regiment shortcut.
+//   3. `stalledZeroRegimentGpPeaceTargets` filters minors and tribes
+//      from `threats.atWarWith` so only Great Powers appear in the
+//      returned list; the companion `stalledZeroRegimentAllFactionPeaceTargets`
+//      owns the minor/tribe arm.
+//   4. `stalledZeroRegimentGpPeaceTargets` sorts the GP list ascending
+//      so the downstream offer-peace pass observes a stable order
+//      regardless of the iteration order of `threats.atWarWith`.
+//   5. `mutualZeroRegimentGpStalematePeaceTargets` short-circuits to
+//      `const []` when the active player's `oldWorldProvincesOwned`
+//      is outside the stalled band, when the active player still has
+//      at least one standing regiment, when the GP-war set is empty,
+//      when the GP-war set has 2+ entries (multi-front shape handled
+//      by `multiFrontNonBlockerGpPeaceTargets`), or when the sole GP
+//      enemy still has at least one standing regiment.
+//   6. `mutualZeroRegimentGpStalematePeaceTargets` returns the
+//      single-element list with the lone GP enemy's `factionId` when
+//      all guards pass.
+//   7. The delegating stubs in `diplomacy_planner_peace_targets.dart`
+//      return the same value as the canonical helpers for every
+//      representative input — required so the legacy fixtures and the
+//      in-file consumer paths agree on the deciders.
+
+import 'package:colonizethis_data/colonizethis_data.dart';
+import 'package:colonizethis_test/test.dart';
+import '../support/expand_phase_peace_test_support.dart';
+import 'package:colonizethis_ai/src/planning/expand_phase_planner.dart';
+
+const String _gpEnemy = 'gp_enemy';
+const String _gpThird = 'gp_third';
+const String _minor1 = 'minor1';
+const String _tribe1 = 'tribe1';
+
+void registerExpandPhasePlannerSurvivalZeroRegimentPeaceGuardCases() {
+  group('stalledZeroRegimentGpPeaceTargets — canonical outer guards', () {
+    test('returns const [] when ownOw exceeds the stalled band '
+        '(ownOw > kStalledOldWorldProvinceThreshold)', () {
+      final game = buildZeroRegimentExpandPeaceGame(
+        ownProvinces: kStalledOldWorldProvinceThreshold + 1,
+        ownRegimentCount: 0,
+        enemyGpIds: const [_gpEnemy],
+        enemyRegimentCount: 0,
+      );
+      final snapshot = ownSnapshot(
+        oldWorldProvincesOwned: kStalledOldWorldProvinceThreshold + 1,
+        atWarWith: const [_gpEnemy],
+      );
+      expect(
+        stalledZeroRegimentGpPeaceTargets(game: game, snapshot: snapshot),
+        isEmpty,
+        reason:
+            'Above the stalled OW band the canonical helper must NOT '
+            'engage the zero-regiment rebuild-peace arm even with zero '
+            'standing regiments. The broader EXPAND deciders own this '
+            'region of the band and a regression that widened the band '
+            'would peace pressing-quota GPs that still hold rebuild-ready '
+            'frontier wars.',
+      );
+    });
+
+    test(
+      'returns const [] when the active player has at least one regiment',
+      () {
+        final game = buildZeroRegimentExpandPeaceGame(
+          ownProvinces: 6,
+          ownRegimentCount: 1,
+          enemyGpIds: const [_gpEnemy],
+          enemyRegimentCount: 0,
+        );
+        final snapshot = ownSnapshot(
+          oldWorldProvincesOwned: 6,
+          atWarWith: const [_gpEnemy],
+        );
+        expect(
+          stalledZeroRegimentGpPeaceTargets(game: game, snapshot: snapshot),
+          isEmpty,
+          reason:
+              'With at least one standing regiment the planner can still '
+              'press existing GP wars; the zero-regiment shortcut must not '
+              'fire. A regression that flipped `> 0` to `>= 0` would peace '
+              'every GP front at non-zero regiment counts and collapse '
+              'EXPAND pressure.',
+        );
+      },
+    );
+
+    test('returns const [] when no Great Powers are at war', () {
+      final game = buildZeroRegimentExpandPeaceGame(
+        ownProvinces: 6,
+        ownRegimentCount: 0,
+        enemyGpIds: const [],
+        enemyRegimentCount: 0,
+        minorIds: const [_minor1],
+        atWarMinorIds: const [_minor1],
+      );
+      final snapshot = ownSnapshot(
+        oldWorldProvincesOwned: 6,
+        atWarWith: const [_minor1],
+      );
+      expect(
+        stalledZeroRegimentGpPeaceTargets(game: game, snapshot: snapshot),
+        isEmpty,
+        reason:
+            'Only a minor is in threats.atWarWith; `game.playerById(...)` '
+            'filters it out and the canonical helper must return an empty '
+            'list. The minor/tribe peace pivot is owned by the companion '
+            'stalledZeroRegimentAllFactionPeaceTargets — a regression '
+            'leaking the minor into this GP arm would double-count the '
+            'minor across both decider families.',
+      );
+    });
+  });
+
+  group('stalledZeroRegimentGpPeaceTargets — canonical firing path', () {
+    test('peaces all at-war Great Powers at the stalled boundary', () {
+      final game = buildZeroRegimentExpandPeaceGame(
+        ownProvinces: kStalledOldWorldProvinceThreshold,
+        ownRegimentCount: 0,
+        enemyGpIds: const [_gpEnemy],
+        enemyRegimentCount: 0,
+      );
+      final snapshot = ownSnapshot(
+        oldWorldProvincesOwned: kStalledOldWorldProvinceThreshold,
+        atWarWith: const [_gpEnemy],
+      );
+      expect(
+        stalledZeroRegimentGpPeaceTargets(game: game, snapshot: snapshot),
+        [_gpEnemy],
+        reason:
+            'At the stalled boundary (ownOw == '
+            'kStalledOldWorldProvinceThreshold) the canonical helper '
+            'must fire when standing regiments == 0. The `<=` boundary '
+            'belongs inside the stalled band; a regression that flipped '
+            '`<=` to `<` would refuse to peace at the band ceiling where '
+            'rebuild is most critical (seed-42 gp6 plateau).',
+      );
+    });
+
+    test('filters minors and tribes out of the GP-only result', () {
+      final game = buildZeroRegimentExpandPeaceGame(
+        ownProvinces: 6,
+        ownRegimentCount: 0,
+        enemyGpIds: const [_gpEnemy],
+        enemyRegimentCount: 0,
+        minorIds: const [_minor1],
+        tribeIds: const [_tribe1],
+        atWarMinorIds: const [_minor1],
+        atWarTribeIds: const [_tribe1],
+      );
+      final snapshot = ownSnapshot(
+        oldWorldProvincesOwned: 6,
+        atWarWith: const [_minor1, _gpEnemy, _tribe1],
+      );
+      expect(
+        stalledZeroRegimentGpPeaceTargets(game: game, snapshot: snapshot),
+        [_gpEnemy],
+        reason:
+            'Only the Great Power must appear in the canonical result; '
+            'minors and tribes route to the companion '
+            'stalledZeroRegimentAllFactionPeaceTargets arm. A regression '
+            'that dropped the `game.playerById(...) != null` filter would '
+            'leak minors and tribes into the GP peace family.',
+      );
+    });
+
+    test('sorts multiple Great Power enemies ascending regardless of '
+        'atWarWith iteration order', () {
+      final game = buildZeroRegimentExpandPeaceGame(
+        ownProvinces: 6,
+        ownRegimentCount: 0,
+        enemyGpIds: const [_gpThird, _gpEnemy],
+        enemyRegimentCount: 0,
+      );
+      final snapshot = ownSnapshot(
+        oldWorldProvincesOwned: 6,
+        atWarWith: const [_gpThird, _gpEnemy],
+      );
+      expect(
+        stalledZeroRegimentGpPeaceTargets(game: game, snapshot: snapshot),
+        [_gpEnemy, _gpThird],
+        reason:
+            'The canonical helper must `..sort()` the GP list so the '
+            'downstream offer-peace pass observes a stable order '
+            'regardless of the iteration order of threats.atWarWith. A '
+            'regression that omitted the sort would leak iteration '
+            'order into the diplomacy pass and break Refs #2509 '
+            "Must-have #7 (determinism).",
+      );
+    });
+  });
+}
