+++ packages/colonizethis_ai/test/planning/expand_phase_planner_survival_zero_regiment_peace_tail_cases.dart
@@ -1,0 +1,298 @@
+// Case bodies for `expand_phase_planner_survival_multi_front_peace_test.dart` (Refs #3977 Phase 6).
+// Registered from the thin contract file of the same stem.
+// Pin/row coverage is preserved 1:1 from the former inline suite.
+
+// Pins canonical homes in `expand_phase_planner.dart` for
+// `stalledZeroRegimentAllFactionPeaceTargets`,
+// `stalledZeroRegimentGpPeaceTargets`,
+// `mutualZeroRegimentGpStalematePeaceTargets`,
+// `mutualExhaustedBelowQuotaGpStalematePeaceTargets`, and
+// `multiFrontNonBlockerGpPeaceTargets` (Refs #2509 S1). Also covers
+// `stalledZeroRegimentGpPeaceTargets` and
+// `mutualZeroRegimentGpStalematePeaceTargets` EXPAND-phase zero-regiment
+// survival peace deciders at their new home in `expand_phase_planner.dart`
+// (Refs #2509 S1).
+//
+// Both deciders were relocated from `diplomacy_planner_peace_targets.dart`
+// so they survive the now-completed S1 deletion of that file. The canonical
+// implementations live in `expand_phase_planner.dart`;
+// `diplomacy_planner_peace_targets.dart` previously retained thin delegating stubs
+// for the legacy `diplomacy_planner_below_quota_peace_part3_test.dart`
+// § "all GP wars when stalled" fixture and the in-file
+// `_survivalGreatPowerPeaceTargets` /
+// `collectStalledGreatPowerPeaceTargets` `zeroRegimentBlockerPeace` /
+// `stalledOwExpansionNeedsPeacePass` consumer chains until the planned
+// deletion.
+//
+// Live consumers (post-relocation):
+//   * `stalledZeroRegimentGpPeaceTargets` is the broad EXPAND
+//     zero-regiment rebuild shortcut from
+//     `SPEC/ai/ai-architecture.md` § Diplomacy targeting — "when
+//     stalled below quota with zero regiments, peace every at-war
+//     Great Power so rebuild is not blocked by futile fronts". It
+//     peaces every at-war GP once the active player is inside the
+//     stalled OW band and holds zero standing regiments, sorted
+//     ascending by `factionId` for downstream offer-peace
+//     determinism.
+//   * `mutualZeroRegimentGpStalematePeaceTargets` is the sole-GP
+//     mutual-stalemate carve-out. It peaces the lone GP enemy when
+//     both sides have zero standing regiments and the active player
+//     is stalled — the only path that exits an army-exhausted GP-only
+//     frontier where the broader `stalledZeroRegimentGpPeaceTargets`
+//     arm is overridden by the `collectStalledGreatPowerPeaceTargets`
+//     GP-only-frontier carve-out re-adding the canonical OW frontier
+//     blocker to the keep-at-war set.
+//
+// Sibling test coverage that this file complements (but does not duplicate):
+//
+//   * `diplomacy_planner_below_quota_peace_part3_test.dart` § "all GP
+//     wars when stalled" exercises the broader arm through the legacy
+//     `diplomacy_planner_peace_targets.dart` entry point and pins the
+//     "filters minors out of GP results" invariant. Both legacy
+//     fixtures depend on the delegating stubs and continue to pass
+//     unchanged after the canonical bodies relocated here.
+//   * `domain_planner_orchestrator_*_two_gp_peace_test.dart` exercise
+//     the deciders through the orchestrator's `runDiplomacyPlanner`
+//     pass under EXPAND / COLONIAL / DEVELOP phases. Those flows rely
+//     on the same post-delegation return values pinned here.
+//
+// Behavioral invariants pinned at the canonical entry points:
+//
+//   1. `stalledZeroRegimentGpPeaceTargets` short-circuits to `const []`
+//      when the active player's `oldWorldProvincesOwned` exceeds
+//      `kStalledOldWorldProvinceThreshold` — outside the stalled band
+//      the rebuild-peace arm does not engage.
+//   2. `stalledZeroRegimentGpPeaceTargets` short-circuits to `const []`
+//      when the active player still holds at least one standing
+//      regiment — the rebuild-peace arm is a zero-regiment shortcut.
+//   3. `stalledZeroRegimentGpPeaceTargets` filters minors and tribes
+//      from `threats.atWarWith` so only Great Powers appear in the
+//      returned list; the companion `stalledZeroRegimentAllFactionPeaceTargets`
+//      owns the minor/tribe arm.
+//   4. `stalledZeroRegimentGpPeaceTargets` sorts the GP list ascending
+//      so the downstream offer-peace pass observes a stable order
+//      regardless of the iteration order of `threats.atWarWith`.
+//   5. `mutualZeroRegimentGpStalematePeaceTargets` short-circuits to
+//      `const []` when the active player's `oldWorldProvincesOwned`
+//      is outside the stalled band, when the active player still has
+//      at least one standing regiment, when the GP-war set is empty,
+//      when the GP-war set has 2+ entries (multi-front shape handled
+//      by `multiFrontNonBlockerGpPeaceTargets`), or when the sole GP
+//      enemy still has at least one standing regiment.
+//   6. `mutualZeroRegimentGpStalematePeaceTargets` returns the
+//      single-element list with the lone GP enemy's `factionId` when
+//      all guards pass.
+//   7. The delegating stubs in `diplomacy_planner_peace_targets.dart`
+//      return the same value as the canonical helpers for every
+//      representative input — required so the legacy fixtures and the
+//      in-file consumer paths agree on the deciders.
+
+import 'package:colonizethis_data/colonizethis_data.dart';
+import 'package:colonizethis_test/test.dart';
+import '../support/expand_phase_peace_test_support.dart';
+import 'package:colonizethis_ai/src/planning/expand_phase_planner.dart';
+
+const String _gpEnemy = 'gp_enemy';
+const String _gpThird = 'gp_third';
+const String _minor1 = 'minor1';
+const String _tribe1 = 'tribe1';
+
+void registerExpandPhasePlannerSurvivalZeroRegimentPeaceTailCases() {
+  group(
+    'mutualZeroRegimentGpStalematePeaceTargets — canonical outer guards',
+    () {
+      test('returns const [] when ownOw exceeds the stalled band '
+          '(ownOw > kStalledOldWorldProvinceThreshold)', () {
+        final game = buildZeroRegimentExpandPeaceGame(
+          ownProvinces: kStalledOldWorldProvinceThreshold + 1,
+          ownRegimentCount: 0,
+          enemyGpIds: const [_gpEnemy],
+          enemyRegimentCount: 0,
+        );
+        final snapshot = ownSnapshot(
+          oldWorldProvincesOwned: kStalledOldWorldProvinceThreshold + 1,
+          atWarWith: const [_gpEnemy],
+        );
+        expect(
+          mutualZeroRegimentGpStalematePeaceTargets(
+            game: game,
+            snapshot: snapshot,
+          ),
+          isEmpty,
+          reason:
+              'Outside the stalled band the mutual-stalemate reset must '
+              'not engage even when both sides hold zero regiments. The '
+              'broader EXPAND consolidate-gains decider owns the '
+              'post-stalled band.',
+        );
+      });
+
+      test(
+        'returns const [] when the active player has at least one regiment',
+        () {
+          final game = buildZeroRegimentExpandPeaceGame(
+            ownProvinces: 6,
+            ownRegimentCount: 1,
+            enemyGpIds: const [_gpEnemy],
+            enemyRegimentCount: 0,
+          );
+          final snapshot = ownSnapshot(
+            oldWorldProvincesOwned: 6,
+            atWarWith: const [_gpEnemy],
+          );
+          expect(
+            mutualZeroRegimentGpStalematePeaceTargets(
+              game: game,
+              snapshot: snapshot,
+            ),
+            isEmpty,
+            reason:
+                'A non-zero standing regiment on the active player means '
+                'the planner can still press the war; the mutual-stalemate '
+                'reset is exclusively a zero-regiment carve-out.',
+          );
+        },
+      );
+
+      test('returns const [] when the sole at-war GP still has regiments', () {
+        final game = buildZeroRegimentExpandPeaceGame(
+          ownProvinces: 6,
+          ownRegimentCount: 0,
+          enemyGpIds: const [_gpEnemy],
+          enemyRegimentCount: 2,
+        );
+        final snapshot = ownSnapshot(
+          oldWorldProvincesOwned: 6,
+          atWarWith: const [_gpEnemy],
+        );
+        expect(
+          mutualZeroRegimentGpStalematePeaceTargets(
+            game: game,
+            snapshot: snapshot,
+          ),
+          isEmpty,
+          reason:
+              'The mutual stalemate requires BOTH sides to be exhausted. '
+              'Pins the enemy-regiment gate against a regression that '
+              'collapsed the `mutualZeroRegimentGpStalemate` carve-out '
+              'into the broader `stalledZeroRegimentGpPeaceTargets` arm.',
+        );
+      });
+
+      test('returns const [] when no Great Power is at war', () {
+        final game = buildZeroRegimentExpandPeaceGame(
+          ownProvinces: 6,
+          ownRegimentCount: 0,
+          enemyGpIds: const [],
+          enemyRegimentCount: 0,
+          minorIds: const [_minor1],
+          atWarMinorIds: const [_minor1],
+        );
+        final snapshot = ownSnapshot(
+          oldWorldProvincesOwned: 6,
+          atWarWith: const [_minor1],
+        );
+        expect(
+          mutualZeroRegimentGpStalematePeaceTargets(
+            game: game,
+            snapshot: snapshot,
+          ),
+          isEmpty,
+          reason:
+              'Zero GP wars cannot return a peace target — only minors / '
+              'tribes are at war here so the canonical helper short-'
+              'circuits without selecting a peace partner.',
+        );
+      });
+
+      test('returns const [] when multiple Great Powers are at war', () {
+        final game = buildZeroRegimentExpandPeaceGame(
+          ownProvinces: 6,
+          ownRegimentCount: 0,
+          enemyGpIds: const [_gpEnemy, _gpThird],
+          enemyRegimentCount: 0,
+        );
+        final snapshot = ownSnapshot(
+          oldWorldProvincesOwned: 6,
+          atWarWith: const [_gpEnemy, _gpThird],
+        );
+        expect(
+          mutualZeroRegimentGpStalematePeaceTargets(
+            game: game,
+            snapshot: snapshot,
+          ),
+          isEmpty,
+          reason:
+              'Multi-front shape (2+ GPs at war) is handled by '
+              'multiFrontNonBlockerGpPeaceTargets — the sole-GP '
+              'mutual-stalemate reset does not apply. A regression that '
+              'dropped the `gpWars.length != 1` guard would double-peace '
+              'across both decider families.',
+        );
+      });
+    },
+  );
+
+  group(
+    'mutualZeroRegimentGpStalematePeaceTargets — canonical firing path',
+    () {
+      test('peaces the sole GP enemy when both sides are exhausted', () {
+        final game = buildZeroRegimentExpandPeaceGame(
+          ownProvinces: kStalledOldWorldProvinceThreshold,
+          ownRegimentCount: 0,
+          enemyGpIds: const [_gpEnemy],
+          enemyRegimentCount: 0,
+        );
+        final snapshot = ownSnapshot(
+          oldWorldProvincesOwned: kStalledOldWorldProvinceThreshold,
+          atWarWith: const [_gpEnemy],
+        );
+        expect(
+          mutualZeroRegimentGpStalematePeaceTargets(
+            game: game,
+            snapshot: snapshot,
+          ),
+          [_gpEnemy],
+          reason:
+              'Both guards (own and enemy regiments == 0) + stalled '
+              'band + exactly one GP war must peace the lone enemy. '
+              'Pins the firing path so the carve-out can never be '
+              'silently retired by an outer-guard refactor on the '
+              'broader stalledZeroRegimentGpPeaceTargets arm.',
+        );
+      });
+
+      test('still peaces when minors are also at war (GP-only filter keeps '
+          'the carve-out tight on the lone GP enemy)', () {
+        final game = buildZeroRegimentExpandPeaceGame(
+          ownProvinces: kStalledOldWorldProvinceThreshold,
+          ownRegimentCount: 0,
+          enemyGpIds: const [_gpEnemy],
+          enemyRegimentCount: 0,
+          minorIds: const [_minor1],
+          atWarMinorIds: const [_minor1],
+        );
+        final snapshot = ownSnapshot(
+          oldWorldProvincesOwned: kStalledOldWorldProvinceThreshold,
+          atWarWith: const [_minor1, _gpEnemy],
+        );
+        expect(
+          mutualZeroRegimentGpStalematePeaceTargets(
+            game: game,
+            snapshot: snapshot,
+          ),
+          [_gpEnemy],
+          reason:
+              'The mutual-stalemate carve-out filters minors out of the '
+              'GP-war set, so a co-belligerent minor at war does not '
+              'switch the helper to the multi-front guard. A regression '
+              'that counted minors in `gpWars.length` would silently '
+              'abandon zero-regiment GPs trapped on mixed-frontier wars.',
+        );
+      });
+    },
+  );
+
+  group('Determinism (Must-have #7)', () {
+  });
+}
