+++ packages/colonizethis_ai/test/planning/phase_planner_conquest_frontier_march_stalled_frontier_cases.dart
@@ -1,0 +1,338 @@
+// Case bodies for `phase_planner_conquest_frontier_march_test.dart` (Refs #4079 Slice D).
+// Registered from the thin contract; pin coverage preserved 1:1 from the
+// former inline suite.
+
+import 'package:colonizethis_ai/colonizethis_ai.dart';
+import 'package:colonizethis_ai/src/planning/conquest_planner.dart';
+import 'package:colonizethis_data/colonizethis_data.dart';
+import 'package:colonizethis_models/colonizethis_models.dart';
+import 'package:colonizethis_test/test.dart';
+
+import '../support/domain_planner_test_fake_api.dart';
+import '../support/planner_test_helpers.dart';
+import 'phase_planner_conquest_frontier_march_support.dart';
+
+void registerPhasePlannerConquestFrontierMarchStalledFrontierCases() {
+  group('runConquestArmyMovePlanner stalled-expansion own-territory '
+      'frontier-march (Refs #2509 EXPAND)', () {
+    test(
+      'EXPAND ow<10 with own-territory-only candidates emits a frontier '
+      'march to the at-war-minor frontier province',
+      () {
+        // Topology: own provinces gp1_inner and gp1_frontier; gp1_frontier
+        // shares an edge with the at-war minor1_a invadable. The army sits
+        // at gp1_inner; its direct neighbors (the suggested destinations
+        // below) are own territory only.
+        const topology = MapTopology(
+          nodes: [
+            TopologyNode(
+              id: 'oldWorld|gp1_inner',
+              regionId: 'oldWorld',
+              type: TopologyNodeType.province,
+            ),
+            TopologyNode(
+              id: 'oldWorld|gp1_frontier',
+              regionId: 'oldWorld',
+              type: TopologyNodeType.province,
+            ),
+            TopologyNode(
+              id: 'oldWorld|gp1_idle',
+              regionId: 'oldWorld',
+              type: TopologyNodeType.province,
+            ),
+            TopologyNode(
+              id: 'oldWorld|minor1_a',
+              regionId: 'oldWorld',
+              type: TopologyNodeType.province,
+            ),
+          ],
+          edges: [
+            TopologyEdge(
+              id1: 'oldWorld|gp1_inner',
+              id2: 'oldWorld|gp1_frontier',
+            ),
+            TopologyEdge(
+              id1: 'oldWorld|gp1_inner',
+              id2: 'oldWorld|gp1_idle',
+            ),
+            TopologyEdge(
+              id1: 'oldWorld|gp1_frontier',
+              id2: 'oldWorld|minor1_a',
+            ),
+          ],
+        );
+        final game = Game(
+          id: 'g-stalled-frontier-march',
+          worldState: WorldState(
+            turnState: const TurnState(
+              phase: TurnPhase.orders,
+              turnNumber: 40,
+            ),
+            oldWorld: RegionData(
+              provinces: const [
+                Province(
+                  id: 'oldWorld|gp1_inner',
+                  regionId: 'oldWorld',
+                  ownerId: 'gp1',
+                ),
+                Province(
+                  id: 'oldWorld|gp1_frontier',
+                  regionId: 'oldWorld',
+                  ownerId: 'gp1',
+                ),
+                Province(
+                  id: 'oldWorld|gp1_idle',
+                  regionId: 'oldWorld',
+                  ownerId: 'gp1',
+                ),
+                Province(
+                  id: 'oldWorld|minor1_a',
+                  regionId: 'oldWorld',
+                  ownerId: 'minor1',
+                ),
+              ],
+            ),
+            newWorld: const RegionData(),
+            armies: const [
+              Army(
+                id: 'army_inner',
+                ownerId: 'gp1',
+                regionId: 'oldWorld',
+                stationedProvinceId: 'oldWorld|gp1_inner',
+                isHomeArmy: false,
+                regimentUnitIds: ['reg1', 'reg2'],
+              ),
+            ],
+          ),
+          players: const [
+            Player(id: 'gp1', displayName: 'P1', isHuman: false),
+          ],
+          minorNations: const [
+            MinorNation(id: 'minor1', displayName: 'Minor 1'),
+          ],
+          aiControlByGpId: const {'gp1': true},
+          diplomacyRelations: const [
+            DiplomacyRelation(
+              factionId1: 'gp1',
+              factionId2: 'minor1',
+              state: RelationState.atWar,
+            ),
+          ],
+        );
+
+        // Suggestion API returns own-territory-only candidates — mirrors the
+        // seed-42 gp1 observed behaviour where the capital army produces 12
+        // suggestions all landing on gp1-owned provinces, none on the
+        // invadable minor1 province.
+        final ctx = buildTestPlannerContext(
+          game: game,
+          topology: topology,
+          nationId: 'gp1',
+          primaryGoal: StrategicGoal.conquer,
+          suggestionAPI: const FakeOrderSuggestionAPIForDomainPlannerTests(
+            work: [],
+            build: [],
+            move: [],
+            research: [],
+            navalMove: [],
+            navalMission: [],
+            diplomatic: [],
+            armyMove: [
+              ArmyMoveOrder(
+                armyId: 'army_inner',
+                destinationProvinceId: 'oldWorld|gp1_frontier',
+              ),
+              ArmyMoveOrder(
+                armyId: 'army_inner',
+                destinationProvinceId: 'oldWorld|gp1_idle',
+              ),
+            ],
+          ),
+        );
+        final snapshot = AIWorldSnapshot(
+          playerId: 'gp1',
+          threats: const ThreatSummary(atWarWith: ['minor1']),
+          opportunities: const OpportunitySummary(),
+          // ow=7 -> stalledExpansion = true (below kStalledOldWorldProvinceThreshold)
+          conquest: const ConquestSummary(
+            oldWorldProvincesOwned: 7,
+            invadableProvinceIdsSorted: ['oldWorld|minor1_a'],
+            adjacentOwnerFactionIdsSorted: ['minor1'],
+          ),
+          colonial: const ColonialSummary(),
+          economy: const EconomySummary(),
+          relations: const {},
+        );
+        const phasePlan = PhasePlanOutcome(
+          phase: ObserverGoalPhase.expand,
+          expandMilitaryPlan: kFrontierMarchExpandOwOnly,
+        );
+
+        final orders = runConquestArmyMovePlanner(
+          ctx: ctx,
+          snapshot: snapshot,
+          declaredWarTargetFactionId: 'minor1',
+          phasePlan: phasePlan,
+        );
+
+        final moves = orders.armyMoveOrdersByPlayerId['gp1'] ?? const [];
+        expect(
+          moves,
+          hasLength(1),
+          reason:
+              'Stalled-expansion EXPAND with own-territory-only candidates '
+              'must still emit one army move (frontier-march) — pre-fix '
+              'returned zero moves and the army parked at capital.',
+        );
+        expect(
+          moves.single.destinationProvinceId,
+          'oldWorld|gp1_frontier',
+          reason:
+              'Frontier-march scoring (`_stalledExpansionArmyMoveScoreDelta` '
+              '→ `_isOnAtWarMinorOrTribeFrontier` + '
+              '`kConquestArmyMoveAdjacentAtWarFrontierBonus`) must prefer '
+              'gp1_frontier (adjacent to at-war minor1) over gp1_idle.',
+        );
+      },
+    );
+
+    test(
+      'EXPAND ow<10 keeps own-territory frontier-march even when no direct '
+      'invadable neighbour exists for any field army',
+      () {
+        const topology = MapTopology(
+          nodes: [
+            TopologyNode(
+              id: 'oldWorld|gp1_capital',
+              regionId: 'oldWorld',
+              type: TopologyNodeType.province,
+            ),
+            TopologyNode(
+              id: 'oldWorld|gp1_frontier',
+              regionId: 'oldWorld',
+              type: TopologyNodeType.province,
+            ),
+            TopologyNode(
+              id: 'oldWorld|minor1_a',
+              regionId: 'oldWorld',
+              type: TopologyNodeType.province,
+            ),
+          ],
+          edges: [
+            TopologyEdge(
+              id1: 'oldWorld|gp1_capital',
+              id2: 'oldWorld|gp1_frontier',
+            ),
+            TopologyEdge(
+              id1: 'oldWorld|gp1_frontier',
+              id2: 'oldWorld|minor1_a',
+            ),
+          ],
+        );
+        final game = Game(
+          id: 'g-stalled-frontier-no-direct',
+          worldState: WorldState(
+            turnState: const TurnState(
+              phase: TurnPhase.orders,
+              turnNumber: 40,
+            ),
+            oldWorld: RegionData(
+              provinces: const [
+                Province(
+                  id: 'oldWorld|gp1_capital',
+                  regionId: 'oldWorld',
+                  ownerId: 'gp1',
+                ),
+                Province(
+                  id: 'oldWorld|gp1_frontier',
+                  regionId: 'oldWorld',
+                  ownerId: 'gp1',
+                ),
+                Province(
+                  id: 'oldWorld|minor1_a',
+                  regionId: 'oldWorld',
+                  ownerId: 'minor1',
+                ),
+              ],
+            ),
+            newWorld: const RegionData(),
+            armies: const [
+              Army(
+                id: 'army_cap',
+                ownerId: 'gp1',
+                regionId: 'oldWorld',
+                stationedProvinceId: 'oldWorld|gp1_capital',
+                isHomeArmy: false,
+                regimentUnitIds: ['reg1'],
+              ),
+            ],
+          ),
+          players: const [
+            Player(id: 'gp1', displayName: 'P1', isHuman: false),
+          ],
+          minorNations: const [
+            MinorNation(id: 'minor1', displayName: 'Minor 1'),
+          ],
+          aiControlByGpId: const {'gp1': true},
+          diplomacyRelations: const [
+            DiplomacyRelation(
+              factionId1: 'gp1',
+              factionId2: 'minor1',
+              state: RelationState.atWar,
+            ),
+          ],
+        );
+        final ctx = buildTestPlannerContext(
+          game: game,
+          topology: topology,
+          nationId: 'gp1',
+          primaryGoal: StrategicGoal.conquer,
+          suggestionAPI: const FakeOrderSuggestionAPIForDomainPlannerTests(
+            work: [],
+            build: [],
+            move: [],
+            research: [],
+            navalMove: [],
+            navalMission: [],
+            diplomatic: [],
+            armyMove: [
+              ArmyMoveOrder(
+                armyId: 'army_cap',
+                destinationProvinceId: 'oldWorld|gp1_frontier',
+              ),
+            ],
+          ),
+        );
+        final snapshot = AIWorldSnapshot(
+          playerId: 'gp1',
+          threats: const ThreatSummary(atWarWith: ['minor1']),
+          opportunities: const OpportunitySummary(),
+          conquest: const ConquestSummary(
+            oldWorldProvincesOwned: 7,
+            invadableProvinceIdsSorted: ['oldWorld|minor1_a'],
+            adjacentOwnerFactionIdsSorted: ['minor1'],
+          ),
+          colonial: const ColonialSummary(),
+          economy: const EconomySummary(),
+          relations: const {},
+        );
+        const phasePlan = PhasePlanOutcome(
+          phase: ObserverGoalPhase.expand,
+          expandMilitaryPlan: kFrontierMarchExpandOwOnly,
+        );
+
+        final orders = runConquestArmyMovePlanner(
+          ctx: ctx,
+          snapshot: snapshot,
+          declaredWarTargetFactionId: 'minor1',
+          phasePlan: phasePlan,
+        );
+
+        final moves = orders.armyMoveOrdersByPlayerId['gp1'] ?? const [];
+        expect(moves, hasLength(1));
+        expect(moves.single.destinationProvinceId, 'oldWorld|gp1_frontier');
+      },
+    );
+
+  });
+}
