+++ tool/check_ai_planning_cases_suite_size.dart
@@ -6,8 +6,9 @@ import 'ct_repo_lint_scan_contract.dart';
 
 /// Physical-line ceiling for AI planning `*_cases.dart` modules
 /// (Refs #3997). Phase-10 Slice C densify ratchets 650→600 (Refs #4104);
-/// Phase-11 Slice C →550 (#4239); Phase-12 Slice D →500 (Refs #4291).
-const int aiPlanningCasesSuitePhysicalLineCeiling = 500;
+/// Phase-11 Slice C →550 (#4239); Phase-12 Slice D →500 (Refs #4291);
+/// Phase-13 Slice D →450 (Refs #4310).
+const int aiPlanningCasesSuitePhysicalLineCeiling = 450;
 
 /// Repo-relative path prefix for AI planning case libraries.
 const String _aiPlanningCasesPathPrefix =
